Update Lists tests
